+++ core/src/commonMain/kotlin/com/github/jcornaz/kwik/generator/Primitives.kt
@@ -1,6 +1,7 @@
 package com.github.jcornaz.kwik.generator
 
 import com.github.jcornaz.kwik.Generator
+import com.github.jcornaz.kwik.map
 import com.github.jcornaz.kwik.withSamples
 import kotlin.random.nextInt
 import kotlin.random.nextLong
@@ -8,10 +9,10 @@ import kotlin.random.nextLong
 /**
  * Returns a generator of integer, all values being generated between [min] and [max] (inclusive)
  *
- * Few edge cases (namely, -1, 0, 1, [min] and [max]) are generated more often than the other values
+ * Contains the samples: -1, 0, 1, [min] and [max]
  */
 fun Generator.Companion.ints(min: Int = Int.MIN_VALUE, max: Int = Int.MAX_VALUE): Generator<Int> {
-    require(max > min) { "Max must be greater than min but min was $min and max was $max" }
+    require(max >= min) { "Max must be greater than min but min was $min and max was $max" }
 
     val range = min..max
     val samples = listOf(0, 1, -1, min, max).filter { it in range }
@@ -22,10 +23,10 @@ fun Generator.Companion.ints(min: Int = Int.MIN_VALUE, max: Int = Int.MAX_VALUE)
 /**
  * Returns a generator of longs, all values being generated between [min] and [max] (inclusive)
  *
- * Few edge cases (namely, -1, 0, 1, [min] and [max]) are generated more often than the other values
+ * Contains the samples: -1, 0, 1, [min] and [max]
  */
 fun Generator.Companion.longs(min: Long = Long.MIN_VALUE, max: Long = Long.MAX_VALUE): Generator<Long> {
-    require(max > min) { "Max must be greater than min but min was $min and max was $max" }
+    require(max >= min) { "Max must be greater than min but min was $min and max was $max" }
 
     val range = min..max
     val samples = listOf(0, 1, -1, min, max).filter { it in range }
@@ -34,37 +35,36 @@ fun Generator.Companion.longs(min: Long = Long.MIN_VALUE, max: Long = Long.MAX_V
 }
 
 /**
- * Returns a generator of floats, all values being generated between [from] and [until] (exclusive)
+ * Returns a generator of floats, all values being generated between [min] and [max] (inclusive)
  *
- * Few edge cases (namely, -1, 0, 1) are generated more often than the other values
+ * Contains the samples: -1, 0, 1, [min] and [max]
+ *
+ * [min] and [max] must be finite.
  */
 fun Generator.Companion.floats(
-    from: Float = Float.NEGATIVE_INFINITY,
-    until: Float = Float.POSITIVE_INFINITY
-): Generator<Float> {
-    require(until > from) { "Until must be greater than from but from was $from and until was $until" }
-
-    val range = from..until
-    val samples = listOf(0f, 1f, -1f).filter { it in range }
-    val generate = create { it.nextFloat() * (until - from) + from }
-
-    return generate.withSamples(samples)
-}
+    min: Float = -Float.MAX_VALUE,
+    max: Float = Float.MAX_VALUE
+): Generator<Float> = doubles(min.toDouble(), max.toDouble()).map { it.toFloat() }
 
 /**
- * Returns a generator of floats, all values being generated between [from] and [until] (exclusive)
+ * Returns a generator of floats, all values being generated between [min] and [max] (inclusive)
  *
- * Few edge cases (namely, -1, 0, 1) are generated more often than the other values
+ * Contains samples: -1, 0, 1, [min] and [max]
+ *
+ * [min] and [max] must be finite.
  */
 fun Generator.Companion.doubles(
-    from: Double = Double.NEGATIVE_INFINITY,
-    until: Double = Double.POSITIVE_INFINITY
+    min: Double = -Double.MAX_VALUE,
+    max: Double = Double.MAX_VALUE
 ): Generator<Double> {
-    require(until > from) { "Until must be greater than from but from was $from and until was $until" }
+    require(max >= min) { "Until must be greater than from but from was $min and until was $max" }
+    require(min.isFinite() && max.isFinite()) { "from or until was not finite" }
 
-    val range = from..until
-    val samples = listOf(0.0, 1.0, -1.0).filter { it in range }
-    val generator = create { it.nextDouble(from, until) }
+    val range = min..max
+    val samples = listOf(0.0, 1.0, -1.0, min, max).filter { it in range }
+
+    val until = (max + Double.MIN_VALUE).takeIf { it.isFinite() } ?: max
+    val generator = create { it.nextDouble(min, until) }
 
     return generator.withSamples(samples)
 }
@@ -73,3 +73,8 @@ fun Generator.Companion.doubles(
  * Returns a generator of booleans
  */
 fun Generator.Companion.booleans(): Generator<Boolean> = create { it.nextBoolean() }
+
+/**
+ * Returns a generator including [Double.NaN] in the samples
+ */
+fun Generator<Double>.withNaN(): Generator<Double> = withSamples(Double.NaN)