+++ core/src/commonTest/kotlin/com/github/jcornaz/kwik/generator/ListGeneratorTest.kt
@@ -3,6 +3,7 @@ package com.github.jcornaz.kwik.generator
 import com.github.jcornaz.kwik.AbstractGeneratorTest
 import com.github.jcornaz.kwik.Generator
 import kotlin.test.Test
+import kotlin.test.assertEquals
 import kotlin.test.assertTrue
 
 class ListGeneratorTest : AbstractGeneratorTest() {
@@ -18,13 +19,28 @@ class ListGeneratorTest : AbstractGeneratorTest() {
     }
 
     @Test
-    fun emitsEmptyLists() {
-        assertTrue(Generator.lists<Int>().randoms(0).take(200).any { it.isEmpty() })
+    fun samplesContainsEmpty() {
+        assertTrue(Generator.lists<Int>().samples.any { it.isEmpty() })
     }
 
     @Test
-    fun emitsSingletonsLists() {
-        assertTrue(Generator.lists<Int>().randoms(0).take(200).any { it.size == 1 })
+    fun samplesContainsSingletons() {
+        assertTrue(Generator.lists<Int>().samples.any { it.size == 1 })
+    }
+
+    @Test
+    fun noEmptySampleWhenMinSizeIsGreaterThan0() {
+        assertTrue(Generator.lists(Generator.ints(), minSize = 1).samples.none { it.isEmpty() })
+    }
+
+    @Test
+    fun noSingletonSampleWhenMinSizeIsGreaterThan1() {
+        assertTrue(Generator.lists(Generator.ints(), minSize = 2).samples.none { it.size <= 1 })
+    }
+
+    @Test
+    fun bigMinSizeIsPossible() {
+        assertEquals(1000, Generator.lists(Generator.ints(), minSize = 1000).randoms(1).first().size)
     }
 
     @Test