+++ core/src/commonTest/kotlin/com/github/jcornaz/kwik/generator/LongGeneratorTest.kt
@@ -2,6 +2,7 @@ package com.github.jcornaz.kwik.generator
 
 import com.github.jcornaz.kwik.AbstractGeneratorTest
 import com.github.jcornaz.kwik.Generator
+import kotlin.math.min
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
@@ -19,13 +20,16 @@ class LongGeneratorTest : AbstractGeneratorTest() {
 
     @Test
     fun produceInsideGivenRange() {
-        assertTrue(Generator.longs(-8, 14).randoms(42).take(200).all { it >= -8 && it <= 14 })
+        assertTrue(Generator.longs(-8, 14).randoms(0).take(1000).all { it >= -8 && it <= 14 })
     }
 
     @Test
-    fun startsWithEdgeCases() {
-        val edgeCases = Generator.longs().randoms(42).take(5).toSet()
+    fun provideSamples() {
+        assertEquals(setOf(Long.MIN_VALUE, Long.MAX_VALUE, -1, 0, 1), Generator.longs().samples)
+    }
 
-        assertEquals(setOf(Long.MIN_VALUE, Long.MAX_VALUE, -1, 0, 1), edgeCases)
+    @Test
+    fun samplesAreInRange() {
+        assertEquals(setOf(42, Long.MAX_VALUE), Generator.longs(min = 42).samples)
     }
 }