+++ core/src/commonTest/kotlin/com/github/jcornaz/kwik/operator/CombineTest.kt
@@ -1,10 +1,10 @@
 package com.github.jcornaz.kwik.operator
 
-import com.github.jcornaz.kwik.AbstractGeneratorTest
-import com.github.jcornaz.kwik.Generator
-import com.github.jcornaz.kwik.combine
-import com.github.jcornaz.kwik.combineWith
+import com.github.jcornaz.kwik.*
+import com.github.jcornaz.kwik.generator.doubles
+import com.github.jcornaz.kwik.generator.ints
 import kotlin.test.Test
+import kotlin.test.assertEquals
 import kotlin.test.assertTrue
 
 class CombineTest : AbstractGeneratorTest() {
@@ -28,6 +28,27 @@ class CombineTest : AbstractGeneratorTest() {
 
         assertTrue(gen.randoms(123).take(200).count { (a, b) -> a != b } > 150)
     }
+
+    @Test
+    fun combineSamples() {
+        val gen = Generator.combine(
+            Generator.create { it.nextInt() }.withSamples(1, 2),
+            Generator.create { it.nextInt().toString() }.withSamples("one", "two")
+        )
+
+        assertEquals(setOf(1 to "one", 1 to "two", 2 to "one", 2 to "two"), gen.samples)
+    }
+
+    @Test
+    fun randomValuesContainsSamples() {
+        val gen = Generator.combine(
+            Generator.ints(min = 3).withSamples(1, 2),
+            Generator.doubles(min = 3.0).withSamples(1.0, 2.0)
+        )
+
+        assertTrue(gen.randoms(0).take(100).any { (x, y) -> x < 3 && y >= 3 })
+        assertTrue(gen.randoms(0).take(100).any { (x, y) -> y < 3 && x >= 3 })
+    }
 }
 
 class CombineWithTransformTest : AbstractGeneratorTest() {
@@ -46,11 +67,39 @@ class CombineWithTransformTest : AbstractGeneratorTest() {
     @Test
     fun combineDifferentValues() {
         val gen = Generator.combine(
-            Generator.create { it.nextInt() },
-            Generator.create { it.nextInt() }
-        ) { a, b -> a to b }
+            Generator.create { it.nextInt() }.withSamples(1, 2),
+            Generator.create { it.nextDouble() }.withSamples(3.0, 4.0)
+        ) { a, b -> CombinedValues(a, b) }
 
-        assertTrue(gen.randoms(123).take(200).count { (a, b) -> a != b } > 150)
+        assertTrue(gen.randoms(0).take(200).count { (a, b) -> a != b.toInt() } > 150)
+    }
+
+    @Test
+    fun combineSamples() {
+        val gen = Generator.combine(
+            Generator.create { it.nextInt() }.withSamples(1, 2),
+            Generator.create { it.nextDouble() }.withSamples(3.0, 4.0)
+        ) { a, b -> CombinedValues(a, b) }
+
+        assertEquals(
+            setOf(
+                CombinedValues(1, 3.0),
+                CombinedValues(1, 4.0),
+                CombinedValues(2, 3.0),
+                CombinedValues(2, 4.0)
+            ), gen.samples
+        )
+    }
+
+    @Test
+    fun randomValuesContainsSamples() {
+        val gen = Generator.combine(
+            Generator.ints(min = 3).withSamples(1, 2),
+            Generator.doubles(min = 3.0).withSamples(1.0, 2.0)
+        ) { a, b -> CombinedValues(a, b) }
+
+        assertTrue(gen.randoms(0).take(100).any { (x, y) -> x < 3 && y >= 3 })
+        assertTrue(gen.randoms(0).take(100).any { (x, y) -> y < 3 && x >= 3 })
     }
 
     private data class CombinedValues(val x: Int, val y: Double)
@@ -72,9 +121,26 @@ class CombineWithTest : AbstractGeneratorTest() {
 
         assertTrue(gen.randoms(123).take(200).count { (a, b) -> a != b } > 150)
     }
+
+    @Test
+    fun combineSamples() {
+        val gen = Generator.create { it.nextInt() }.withSamples(1, 2)
+            .combineWith(Generator.create { it.nextInt().toString() }.withSamples("one", "two"))
+
+        assertEquals(setOf(1 to "one", 1 to "two", 2 to "one", 2 to "two"), gen.samples)
+    }
+
+    @Test
+    fun randomValuesContainsSamples() {
+        val gen = Generator.ints(min = 3).withSamples(1, 2)
+            .combineWith(Generator.create { it.nextInt().toString() }.withSamples("one", "two"))
+
+        assertTrue(gen.randoms(0).take(100).any { (x, y) -> x < 3 && y !in setOf("one", "two") })
+        assertTrue(gen.randoms(0).take(100).any { (x, y) -> y in setOf("one", "two") && x >= 3 })
+    }
 }
 
-class ZipWithTransformTest : AbstractGeneratorTest() {
+class CombineWithWithTransformTest : AbstractGeneratorTest() {
 
     override val generator: Generator<*> =
         Generator.create { it.nextInt() }
@@ -91,6 +157,34 @@ class ZipWithTransformTest : AbstractGeneratorTest() {
             .combineWith(Generator.create { it.nextInt() }) { a, b -> a to b }
 
         assertTrue(gen.randoms(123).take(200).count { (a, b) -> a != b } > 150)
+    }
+
+    @Test
+    fun combineSamples() {
+        val gen = Generator.create { it.nextInt() }.withSamples(1, 2)
+            .combineWith(Generator.create { it.nextDouble() }.withSamples(3.0, 4.0)) { a, b ->
+                CombinedValues(a, b)
+            }
+
+        assertEquals(
+            setOf(
+                CombinedValues(1, 3.0),
+                CombinedValues(1, 4.0),
+                CombinedValues(2, 3.0),
+                CombinedValues(2, 4.0)
+            ), gen.samples
+        )
+    }
+
+    @Test
+    fun randomValuesContainsSamples() {
+        val gen = Generator.ints(min = 3).withSamples(1, 2)
+            .combineWith(Generator.doubles(min = 3.0).withSamples(-1.0, -2.0)) { a, b ->
+                CombinedValues(a, b)
+            }
+
+        assertTrue(gen.randoms(0).take(100).any { (x, y) -> x < 3 && y >= 3 })
+        assertTrue(gen.randoms(0).take(100).any { (x, y) -> y < 3 && x >= 3 })
     }
 
     private data class CombinedValues(val x: Int, val y: Double)