+++ travis_script.sh
@@ -1,10 +1,10 @@
 #!/bin/bash
 
 
-if [ "$TRAVIS_BRANCH" = "master" ] && [ "$TRAVIS_PULL_REQUEST" = "false" ]
+if [ "$TRAVIS_PULL_REQUEST" = "false" ]
 then
     ./gradlew check bintrayUpload sonarqube --no-daemon --no-build-cache
-elif [ "$TRAVIS_PULL_REQUEST" = "false" ] || [ -z "$SONAR_TOKEN" ]
+elif [ -z "$SONAR_TOKEN" ]
 then
     ./gradlew check
 else